login function working, register new user working, delete/edit post only available for user/admin that created post.

+++ app/views/page/show.php
@@ -1,5 +1,4 @@
 <?php require VIEW_ROOT . '/templates/header.php'; ?>
-
   <?php if(!$data): ?>
     <p>No page found, sorry.</p>
   <?php else: ?>
@@ -14,7 +13,15 @@
       <?php endif; ?>
     </p>
   <?php endif; ?>
-  <a href="<?php echo BASE_URL; ?>/edit_post.php?id=<?php echo $data['post_id']; ?>">Edit post</a>
-  <a href="<?php echo BASE_URL; ?>/delete_post.php?id=<?php echo $data['post_id']; ?>">Delete post</a>
+
+
+  <?php if($_SESSION == true): ?>
+    <?php if($_SESSION['user_id'] == $data['user_id'] OR $_SESSION['admin'] == 1): ?>
+      <a href="<?php echo BASE_URL; ?>/delete_post.php?id=<?php echo $data['post_id']; ?>">Delete post</a>
+      <?php if($_SESSION['user_id'] == $data['user_id']): ?>
+        <a href="<?php echo BASE_URL; ?>/edit_post.php?id=<?php echo $data['post_id']; ?>">Edit post</a>
+      <?php endif; ?>
+    <?php endif; ?>
+  <?php endif; ?>
 
 <?php require VIEW_ROOT . '/templates/footer.php'; ?>
